Update architecture to clean arch
